Move github client to better folder
