+++ src/commands/whoami.ts
@@ -1,8 +1,8 @@
 import { Command } from "@oclif/command"
 import cli from "cli-ux"
 import fetch from "node-fetch"
-import { Config } from "../config"
 import { Gravity } from "../clients/gravity"
+import { Config } from "../config"
 
 export default class WhoAmI extends Command {
   static description = "Who are you?"
